Created schedular for save template analytics data

+++ src/F4ConversationCloud/F4ConversationCloud.Application/Common/Models/TemplateAnalyticsDTO.cs
@@ -21,6 +21,6 @@ namespace F4ConversationCloud.Application.Common.Models
         public DateTime TemplateSendTo { get; set; }
         public string WABAID { get; set; }
         public int ClientInfoId { get; set; }
-        public int TemplateId { get; set; }
+        public long TemplateId { get; set; }
     }
 }
